Add checking user already moveseat and change user identifier email to code

+++ endpoint/api/move-seat/group.js
@@ -4,28 +4,32 @@ const isNumber = require('is-number');
 
 module.exports = {
     post: async (ctx, next) => {
+        if(!isNumber(ctx.request.body.classCode) || ctx.request.body.classCode <= 0) ctx.throw(400);
         const classCode = parseInt(ctx.request.body.classCode, 10);
         const classInfo = await ctx.state.collection.classes.findOne({ code: classCode });
         if(((1 << (ctx.state.user.grade-1)) & classInfo.moveseat.group) == 0) ctx.throw(400);
         const count = await ctx.state.collection.moveSeatState.countDocuments({ classcode: classCode });
         if(!ctx.request.body.users || !classInfo || count >= 1) ctx.throw(400);
-        
-        const users = JSON.parse(ctx.request.body.users);
-        if(!Array.isArray(users)) ctx.throw(400);
-        if(users.length == 0 || users.length+1 > classInfo.moveseat.limit) ctx.throw(400);
+        if(ctx.state.user.moveSeatInfo.classCode != 0 && isSameDay(ctx.state.user.moveSeatInfo.date, new Date())) ctx.throw(400);
+
+        const usersArray = JSON.parse(ctx.request.body.users);
+        if(!Array.isArray(usersArray)) ctx.throw(400);
+        if(usersArray.length == 0 || usersArray.length+1 > classInfo.moveseat.limit) ctx.throw(400);
 
         const userinfo = await Promise.all(
-            users.map(async (user) => {
+            usersArray.map(async (user) => {
                 if(!isNumber(user)) ctx.throw(400);
                 const ithuser = await ctx.state.collection.users.findOne({ code: user });
                 if(!ithuser) ctx.throw(403);
                 ithuser.password = undefined;
                 if(((1 << (ithuser.grade-1)) & classInfo.moveseat.group) == 0) ctx.throw(400);
+                if(ithuser.moveSeatInfo.classCode != 0 && isSameDay(ithuser.moveSeatInfo.date, new Date())) ctx.throw(400);
+                ithuser.moveSeatInfo = undefined;
                 return ithuser;
             })
         );
 
-        await ctx.state.collection.moveSeatState.findOneAndUpdate({ 'primary.email': ctx.state.user.email },
+        await ctx.state.collection.moveSeatState.findOneAndUpdate({ 'primary.code': parseInt(ctx.state.user.code, 10) },
             { $setOnInsert:
                 {
                     classcode: classCode,
@@ -34,6 +38,12 @@ module.exports = {
                 }
             }, { upsert: true }
         );
+        
+        await ctx.state.collection.users.findOneAndUpdate({ code: parseInt(ctx.state.user.code, 10) }, { $set: { moveSeatInfo: { date: new Date(), classCode: parseInt(classCode, 10), type: "group" } } } );
+        userinfo.forEach(async (user) => {
+            await ctx.state.collection.users.findOneAndUpdate({ code: user.code }, { $set: { moveSeatInfo: { date: new Date(), classCode: parseInt(classCode, 10), type: "group" } } } );
+        });
+
         await next();
     },
     get: async (ctx, next) => {
@@ -41,12 +51,13 @@ module.exports = {
         await next();
     },
     delete: async (ctx, next) => {
-        await ctx.state.collection.moveSeatState.deleteOne({ 'primary.email': ctx.state.user.email });
+        await ctx.state.collection.moveSeatState.deleteOne({ 'primary.code': parseInt(ctx.state.user.code, 10) });
         await next();
     },
     put: async (ctx, next) => {
+        if(!isNumber(ctx.request.body.classCode) || ctx.request.body.classCode <= 0) ctx.throw(400);
         if(!isNumber(ctx.request.body.userInfo) || !isNumber(ctx.request.body.operation)) ctx.throw(400);
-        const moveSeatInfo = await ctx.state.collection.moveSeatState.findOne({ 'primary.email': ctx.state.user.email });
+        const moveSeatInfo = await ctx.state.collection.moveSeatState.findOne({ 'primary.code': parseInt(ctx.state.user.code, 10) });
         if(!moveSeatInfo) ctx.throw(400);
         const userIndex = moveSeatInfo.secondary.findIndex(element => (element.code == parseInt(ctx.request.body.userInfo, 10)));
 
@@ -65,22 +76,16 @@ module.exports = {
             }
         } else ctx.throw(400);
 
-        await ctx.state.collection.moveSeatState.updateOne({ 'primary.email': ctx.state.user.email }, { $set: { secondary: moveSeatInfo.secondary } });
+        await ctx.state.collection.moveSeatState.findOneAndUpdate({ 'primary.code': parseInt(ctx.state.user.code, 10) }, { $set: { secondary: moveSeatInfo.secondary } });
         await next();
     },
     common: async (ctx, next) => {
-        if(!isNumber(ctx.request.body.classCode) || ctx.request.body.classCode <= 0) ctx.throw(400);
         if(!ctx.state.user) ctx.throw(400);
 
         ctx.state.collection.moveSeatState = ctx.state.db.collection(`move-seat-group`);
-        const randomDoc = (await ctx.state.collection.moveSeatState.aggregate([{
-            $sample: {
-                size: 1
-            }
-        }]).toArray())[0];
-        if(!randomDoc || !isSameDay(new ObjectId(randomDoc._id).getTimestamp(), new Date())) {
-            await ctx.state.collection.moveSeatState.deleteMany({});
-        }
+        
+        const randomDoc = (await ctx.state.collection.moveSeatState.aggregate([{ $sample: { size: 1 } }]).toArray())[0];
+        if(!randomDoc || !isSameDay(new ObjectId(randomDoc._id).getTimestamp(), new Date())) { await ctx.state.collection.moveSeatState.deleteMany({}); }
 
         await next();
     }
